Added capability to save png or other static format
As I'm looking to use this Sphinx extension on my module, considering the
high number of plots (many of which are 3D plots that requires a lot of
bandwidth and really slows down the rendering of the page), I figured
that it would be nice to visualize png plots.

This commits modifies `save_plotly_figure` and add a couple of tests to
verify its correct behaviour.

Users can set the following on `conf.py` (for example):

plotly_formats = ["png", "html", "pdf"]

This will visualize png by default, also adding links to a html and pdf
versions.